refactor: replace read-pkg-up with escalade
Ref https://packagephobia.com/result?p=read-pkg-up

read-pkg-up is quite big package with many dependencies.
It solves cases not necessary for this project like pretty json errors
and normalisation of package.json data.

In this diff I replaced it with dependency-free escalade and parsed json
with native JSON.parse.

This will let users to have faster installation and less bloated
node_modules.