fix: error on hmr when file path is null (#213)
* Fixed documentation for Webpack output.library and libraryTarget: 'window'

When building a library and use a multi entry config, only the last item in the entry array will be exposed. The example that was provided by the plugin would expose the webpack plugin serve client in the end file and it breaks the app

* Fixed error breaking hmr

filePath on webpack invalid hook can return filePath or null. This resolves the following error when it returns null

Cannot read property 'replace' of null